Moved tests to test/files/ folder.
